Encapsulate the keys of fladgejt.structures.
We refer to most fladgejt structures by some subset of their fields that we
assume will be immutable and mostly unique. These are used as a key in
find_row(), and later request_json(), to look up the full instance. But most
fladgejt methods don't care about the key's exact composition. Instead of
having the key's fields directly in argument lists, we group them in a tuple.
Converting a full instance to this smaller tuple can be done with a new "key"
field that is automatically initialized on namedtuple creation.

Fixes #12.