Refactor keys of nested structures.
All keys are now globally unique. If structure B "belongs to" structure A,
B.b_key now contains A.a_key. Previously, you needed both a_key and b_key to
uniquely identify an instance of B. This led to some client side bugs where
only b_key was used as a React key.

Keys are always strings. Previously, foo_key was a string on the client and
a decoded tuple on the server. Encoding and decoding happened transparently in
the RPC layer. Now keys must be explicitly decoded. This is mainly to allow
keys to contain other keys, but it should be simpler in general. This can lead
to more encoding and decoding, but the cost should be negligible.

Also fix a bug in odhlas_z_terminu() where predmet_key was completely ignored.

As a slight regession, studiumKey is no longer preserved when switching pages
from "Moje predmety" or "Moje skusky" to "Moje hodnotenia".