Correct errors to user after AIS login expired #76
* When user click on ‘Moje predmety’ after AIS login expired, the
request goes through REST. The server already returning correct error,
but it is necessary to display the error correctly to the user.
In file ErrorPage.js, two types of error are solved. Errors of
votr.settings.error type and errors which comes from AJAX. In both cases
should be given LogOutError condition. That was fixed in this commit.

* When user click on ‘Register osob’after AIS login expired, the request
goes through AIS. In this case AIS returns always three the same
operations with methods (serverCloseApplication, messageBox,
closeApplication). This situation should be detect in Aplication.open
and LogOutError should be throwen after that. It was fixed in this
commit.

* Error message for situations during connection problems. In file
ErrorPage.js I added condition which catch
requests.exceptions.ConnectionError. After that, the user will be
notified about connection issues by connection error message.

Signed-off-by: Lukáš Mayer <luky.mayer@gmail.com>