Add a custom StatusFilePlugin for webpack.
The plugin creates a file that says whether webpack is running, has succeeded,
or has failed. front.py reads the file and waits for the build to finish if
it's currently in progress.

This is essentially a custom substitute for webpack-dev-server. We could use
that instead, but I couldn't find a way to configure it to do what I want.
webpack-dev-server has two modes: inline mode and iframe mode. In inline mode,
it injects its own client library into every bundle. This is a problem for Votr
since we load both prologue.?.js and votr.?.js on every page. In iframe mode,
they are separated into different frames, but the server on being on top and
putting Votr in the iframe, not vice versa. When Votr navigates to another URL
or changes the page title, the user won't see. There might be a third option to
load /webpack-dev-server.js as a separate script in front.py, but it is not
documented anywhere. The documentation is lacking in general.

It's also arguable if we even want automatic refresh after every rebuild, when
AIS requests can take so long. Hot module reloading could solve that, but it's
also more difficult to configure. We can take another look later.