Change a few remaining key= attributes to be always unique.
Element lists in React must have a unique key={...} attribute. This used to be
a big problem for Votr because most tables directly used fladgejt structure
keys, which weren't always unique. It was fixed when we changed all tables to
use SortableTable, because it uses key={originalIndex} on table rows.

This commit just hardens the last few cases to really make sure they're unique.
They are very unlikely to have duplicates anyway.

Updates #91.