Remove hodn_key. It is not used anymore.
FWIW: adding 'semester' is not enough to make it unique. It is possible for a
given student to be enrolled in a given course multiple times, even in the same
year and semester. Usually (maybe always?) they're postgraduate courses like
"Pedagogická činnosť", "Vedecká činnosť III", "Seminár pracoviska" etc. There
are known cases of:

- two rows which only differ in hodn_datum, hodn_termin, hodn_znamka,
  hodn_znamka_popis (empty vs non-empty) (seen with semester = "" or "Z")

- two rows which only differ in hodn_datum, hodn_znamka, hodn_znamka_popis
  (different non-empty values) (hodn_termin is "R - ...", semester is "")

- two rows which only differ in hodn_datum (e.g. "09.02.20xx" vs "19.06.20xx")

- two completely equal rows (seen with semester = "" or "Z")

Fortunately, we don't need a unique row identifier for this table.

Updates #91.