Erase the URL query on login unless the user opts in.
Users often open Votr with a saved link from their bookmarks or history. If the
link is bad (e.g. it has an invalid key for some reason), the user gets an error
every time they log in with it. Even for valid URLs, I'm not so sure whether the
user really wanted to visit that page or it's just a waste of resources.

The login page now has a checkbox to respect the current URL. It is false by
default, but it is persisted in localStorage. I hope this will be better for
users who use saved links by accident, while still allowing power users to
bookmark pages if they want.

Stats for 2023-01:
- 71% of logins had empty destination
- 21% of logins only had ?action= in destination
- 8% of logins had multiple query parameters

Updates #65. Updates #133.