Reduce installed size of eslint-plugin-react.
eslint-plugin-react depends on many polyfills because it supports all versions
since node.js 4. Which is fine, it just makes the node_modules directory larger.
That's not a problem either. But it mildly irritates me.

du -sh node_modules:
  without eslint-plugin-react:   223M
  with eslint-plugin-react:      236M (+13M)
  after this commit:             225M (-11M)

I know this is a silly thing to optimize, but at least I had fun, I guess...?