Add a home link to HTTP error pages.
Most users should never see or need this, but just in case.

One known case is a very infrequent error 405 on /login and /reset. It could be
that the browser can't send the POST because of a bad connection, and when the
user refreshes later, the browser forgets it was supposed to be a POST request.