Batch multiple RPC calls in one /rpc HTTP request.
It works, but it's not as useful as I hoped:

I wanted to stop having wasteful idle gunicorn workers just waiting for a
session lock. And this does help a bit. But there aren't that many batched RPCs
in practice. And if the user navigates away and makes another RPC, the worker
still has to wait.

I hoped this might give us easy analytics of user navigations just from the
server access log, but many pages still send multiple /rpc HTTP requests. It's
far from 1:1.