Migrate JS package management from yarn 1 to pnpm.
yarn 1 is starting to show its age, with deprecation warnings and occasional
random errors. I tried yarn 4, but I'm not a fan:

- `yarn up` is confusing and hard to use (e.g. berry#246, berry#477)
- `yarn install --prod` was replaced by `yarn workspaces focus --production`
- the .gitignore rules needed for .yarn/* are not very pretty

The remaining options are npm and pnpm. Both are good. I chose pnpm for now:

- `pnpm update` has almost exactly the behavior and options I want
- `pnpm patch` looks neat (but isn't used for now)
- the symlink structure of node_modules allows simplifying our tsconfig.json
- using hardlinks and reflinks looks neat (but isn't a huge deal for Votr dev)

pnpm can be installed in many ways. I picked corepack with a pinned version in
package.json. corepack adds a bit of start time overhead, but it's <100ms and
Votr does not need to launch pnpm often.

stats:
  du -sh node_modules
  wc -l yarn.lock pnpm-lock.yaml
  find node_modules -name package.json | wc -l
before: 228M | 2337 | 469
after:  219M | 3009 | 469