Tweak saml_logout logic.
Calling do_logout in both branches is not exactly a bug, but it makes more sense
to call it only for externally initiated logout.

Returning app_response() was fine, but if #163 is fully fixed sometime in the
future, the destination= kwarg might be removed. The response is not shown
anyway -- Shibboleth IdP loads it in a hidden iframe.