Save sessions even on errors, but close all apps afterwards.
Until now, a failed RPC reverted the session to its previous state. I think the
original rationale was that errors can cause anything to happen, so even dumping
and loading the pickle could cause trouble. In practice, it turns out there are
only few kinds of errors, and they mostly behave predictably.

Reverting the session is wrong because it can desynchronize GUI state between
AIS and Votr (Aisikl) when the next RPC executes. If the failed RPC opened an
app, it won't count towards OPEN_APP_LIMIT. If it closed a dialog, `app.d` will
still point to it. If it selected another row of a table, `selection_changed`
may be wrong, causing us to not send a selection event.

But some worry might be warranted. Can all Fladgejt code start from any GUI
state and recover perfectly? One example: If BrokenPipeError happens during
_select_ciselnik() in vyhladaj_osobu(), then the enum dialog "SSSC001" aka
"CiselnikDlg2" will stay open. The next call to vyhladaj_osobu() expects to see
either the main dialog "LZ014_VyhladavanieOsobDlg0" or the filter dialog
"LZ015_VyhladavanieOsobFilterDlg1", it doesn't know about "CiselnikDlg2".

Therefore, if we notice the last RPC failed, we close all open applications.
Starting over is usually the safest option.

As a special case, if the RPC error was a timeout during request_ais(), the AIS
server may reject the WebUIKillEvent and say it's still busy with a previous
request. In this case, we must remember the application is still open, and
adjust the pool limit accordingly.

Closes #62, although not how originally planned. Hopefully, after this change
it's safe to keep using a session after an RPC error.